move registry service to common
