+++ api-node/src/common/registry.service.ts
